feat: update User model to use Float for current_balance, total_expenses, and total_income; adjust transaction handling in routes

+++ db/models.py
@@ -17,12 +17,12 @@ class User(Base):
     name = Column(String)
     email = Column(String)
     password = Column(String)
-    current_balance = Column(Integer, default=0)
+    current_balance = Column(Float, default=0)
     transactions = relationship("Transaction", back_populates="user")
     categories = relationship("Category", back_populates="user")
     total_transactions = Column(Integer, default=0)
-    total_expenses = Column(Integer, default=0)
-    total_income = Column(Integer, default=0)
+    total_expenses = Column(Float, default=0)
+    total_income = Column(Float, default=0)
 
 
 class Transaction(Base):
